% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ai_pubmed}
\alias{ai_pubmed}
\title{Pubmed articles about AI until november 2019}
\format{A data frame with 101,000 rows and 28 variables:
\describe{
  \item{title}{Title}
  \item{year}{Year}
  \item{journal}{Journal}
  \item{pmid}{Pubmed ID}
  \item{keywords}{keywords}
  \item{authors}{authors}
  \item{abstract}{abstract}
  \item{NLMid}{NLMid}
}}
\usage{
ai_pubmed
}
\description{
Pubmed articles about AI until november 2019
}
\keyword{datasets}
