\name{alphatize}
\alias{alphatize}
\title{Apply alpha transparency to a color vector}
\usage{
  alphatize(color, alpha)
}
\arguments{
  \item{color}{A vector of colors}

  \item{alpha}{A vector of alpha values, recycled as
  necessary}
}
\value{
  A vector of colors, with alpha applied
}
\description{
  Apply alpha transparency to a color vector
}
\author{
  David C. Norris
}
\keyword{color}

