% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{base_pop}
\alias{base_pop}
\title{Populations of cities in France 1876-2015.}
\format{A data frame with 1027122 rows and 7 variables:
\describe{
  \item{code_geo}{geographic_code of the city}
  \item{region}{the region of the city}
  \item{dep}{the departement number of the city}
  \item{ville}{the name of the city}
  \item{annee}{the year of the census}
  \item{population}{the population of the city}
  \item{dep_libelle}{the name of the departement of the city}
}}
\source{
\url{https://www.insee.fr/fr/statistiques/fichier/3698339/base-pop-historiques-1876-2015.xls}
}
\usage{
base_pop
}
\description{
A dataset containing the populations of all cities in France based on INSEE census
between 1876 and 2015.
}
\keyword{datasets}
