% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbp_color.R
\name{cbp_pal}
\alias{cbp_pal}
\title{A colorblind proof color palette}
\usage{
cbp_pal(black = FALSE)
}
\description{
A colorblind proof color palette
}
\examples{
scales::show_col(cbp_pal(black = FALSE)(8))
scales::show_col(cbp_pal(black = TRUE)(8))

}
