% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{color}
\alias{color}
\title{Generic function \code{color}}
\usage{
color(x, ...)
}
\arguments{
\item{x}{A colored object}

\item{\dots}{Additional arguments}
}
\value{
A vector of colors
}
\description{
This generic function supports extraction of colors from colored
objects
}

