% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_hex_palette_from_number.R
\name{color_hex_palette_from_name}
\alias{color_hex_palette_from_name}
\title{[color-hex.com](http://color-hex.com) palette by name}
\usage{
color_hex_palette_from_name(name)
}
\arguments{
\item{name}{name of the palette}
}
\value{
character vector of hexadecimal colors
}
\description{
Retrieves a color palette from [color-hex.com](http://color-hex.com) identified by its name.
}
\examples{
color_hex_palette_from_name("Cook")

}
