% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_hex_palette_from_number.R
\name{color_hex_palette_from_number}
\alias{color_hex_palette_from_number}
\title{color-hex.com palette by number}
\usage{
color_hex_palette_from_number(number)
}
\arguments{
\item{number}{palette number from color-hex.com}
}
\value{
vector of hexadecimal colors
}
\description{
Retrieves a color palette from color-hex.com identified by its number as found in the url
of the palette page (e.g. in "https://www.color-hex.com/color-palette/5361", 5361 is the number).
}
\examples{
color_hex_palette_from_number(5361)
}
