% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dataset.R
\name{draw_uniform_sample}
\alias{draw_uniform_sample}
\title{Draw Uniform Sample}
\usage{
draw_uniform_sample(size, min, max, group = 1)
}
\arguments{
\item{size}{number. size of the sample}

\item{min}{number. minimum of the uniform distribution}

\item{max}{number. maximum of the uniform distribution}

\item{group}{number. number of the group (default = 1)}
}
\value{
data.frame
}
\description{
draws a uniform sample of size `size`
}
\examples{
draw_uniform_sample(size = 20, min = 0, max = 5, group = 1)

}
