\name{[.colored}
\alias{[.colored}
\title{Index a colored vector}
\usage{
  \method{[}{colored} (x, ...)
}
\arguments{
  \item{x}{A colored factor}

  \item{...}{Unused}
}
\value{
  A colored factor
}
\description{
  Index a colored vector
}
\author{
  David C. Norris
}
\seealso{
  \code{\link{colored}}
}

