% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grippe}
\alias{grippe}
\title{Sentiweb Incidence Data file for flu (grippe)}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 23764 rows and 9 columns.}
\source{
\url{https://www.sentiweb.fr/datasets/incidence-RDD-3.csv}
}
\usage{
grippe
}
\description{
A dataset containing the incidences of Influenza-like illness by region in France
between 1984 and 2019.
}
\keyword{datasets}
