% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minard_cities}
\alias{minard_cities}
\title{Minard: cities}
\format{A data frame with 20 rows and 3 variables:
\describe{
  \item{long}{Longitude}
  \item{lat}{Latitude}
  \item{city}{City}
}}
\source{
\url{http://www.datavis.ca/gallery/minard/ggplot2/}
}
\usage{
minard_cities
}
\description{
A dataset containing the cities for the Minard dataviz.
}
\keyword{datasets}
