% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minard_temps}
\alias{minard_temps}
\title{Minard: temps}
\format{A data frame with 9 rows and 5 variables:
\describe{
  \item{long}{Longitude}
  \item{temp}{Temperature in °C}
  \item{month}{Month}
  \item{day}{Day}
  \item{date}{Date}
}}
\source{
\url{http://www.datavis.ca/gallery/minard/ggplot2/}
}
\usage{
minard_temps
}
\description{
A dataset containing the temperatures for the Minard dataviz.
}
\keyword{datasets}
