% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minard_troops}
\alias{minard_troops}
\title{Minard: troops}
\format{A data frame with 51 rows and 5 variables:
\describe{
  \item{long}{Longitude}
  \item{lat}{Latitude}
  \item{survivors}{Survivors}
  \item{direction}{Direction A = Aller, R = Retour}
  \item{group}{group}
}}
\source{
\url{http://www.datavis.ca/gallery/minard/ggplot2/}
}
\usage{
minard_troops
}
\description{
A dataset containing the troops for the Minard dataviz.
}
\keyword{datasets}
