% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.marginal.density.R
\name{panel.marginal.density}
\alias{panel.marginal.density}
\title{panel.marginal.density}
\usage{
panel.marginal.density(x, y, darg = list(n = 30), plot.points = FALSE,
  groups = NULL, weights = NULL, jitter.amount = 0.01 *
  diff(current.panel.limits()$ylim), type = "p", ...,
  identifier = "density", intrude = 0.1)
}
\arguments{
\item{x}{}

\item{y}{}

\item{...}{}
}
\value{
None
}
\description{
Panel function for supplementing scatter plots with marginal density graphics
}
\details{
This panel function is adapted directly from lattice::panel.densityplot,
and aims to make inclusion of optional density graphics on either or both
axes of x-y scatter plots.
}
\author{
David C. Norris
}
\seealso{
\code{\link{panel.densityplot}}
}
\keyword{hplot}
\keyword{internal}

