% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_pal.R
\name{show_pal}
\alias{show_pal}
\title{Show Palette}
\usage{
show_pal(pal)
}
\arguments{
\item{pal}{character vector of hexadecimal colors}
}
\value{
plot of pal
}
\description{
Plots a color palette.
}
\examples{
pal <- color_hex_palette_from_number(5361)
show_pal(pal)
}
