% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_elegant.R
\name{theme_elegant}
\alias{theme_elegant}
\title{Theme Elegant}
\usage{
theme_elegant(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{base_size}

\item{base_family}{base_family}

\item{base_line_size}{base_line_size}

\item{base_rect_size}{base_rect_size}
}
\value{

}
\description{
An elegant ggplot theme for modern visualizations
(Basically a copy of `see::theme_modern` but with slight tweeks)
}
\examples{
library(ggplot2)
ggplot(iris, aes(x = Sepal.Width, y = Sepal.Length, color = Species)) +
 geom_point() +
 theme_elegant()

}
