% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_minimal_grid.R
\name{theme_minimal_grid}
\alias{theme_minimal_grid}
\title{Theme Minimal for grid display}
\usage{
theme_minimal_grid(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{base_size}

\item{base_family}{base_family}

\item{base_line_size}{base_line_size}

\item{base_rect_size}{base_rect_size}
}
\value{
ggplot2 theme
}
\description{
A theme compatible with grid display (using Patchwork)
}
\examples{


}
