% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_minimal_grid.R
\name{theme_void_complete}
\alias{theme_void_complete}
\title{Theme completely void (no legend)}
\usage{
theme_void_complete(base_size = 11, base_family = "",
  base_line_size = base_size/170, base_rect_size = base_size/170)
}
\arguments{
\item{base_size}{base_size}

\item{base_family}{base_family}

\item{base_line_size}{base_line_size}

\item{base_rect_size}{base_rect_size}
}
\description{
Theme completely void (no legend)
}
