% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_city_populations}
\alias{us_city_populations}
\title{Populations of big US cities 1790-2010.}
\format{A data frame with 1898 rows and 6 variables:
\describe{
  \item{Rank}{the rank of the city for a given year}
  \item{City}{the name of the city}
  \item{State}{the state of the city}
  \item{Population}{the population of the city}
  \item{Region}{the region of the city among : Midwest, Northeast, South, West }
  \item{Year}{the year of the census}
}}
\usage{
us_city_populations
}
\description{
A dataset containing the populations of big US cities based on US census
between 1790 and 2010.
}
\keyword{datasets}
