% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizOR-package.R
\docType{package}
\name{VizOR-package}
\alias{VizOR}
\alias{VizOR-package}
\title{Graphical visualization tools for complex observational data (VizOR).
Package functionality includes individual- and aggregate-level graphics
depicting patterns of treatment and response in patient registries, and
a graphical tool for identifying potential for confounding in analyses
of observational data.}
\description{
\tabular{ll}{
Package: \tab VizOR\cr
Type: \tab Package\cr
Version: \tab 0.8-3\cr
Date: \tab 2014-09-11\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
TODO: Provide a general introduction to the package, including an orientation
to its most important functions.
}
\author{
Drew Griffin Levy, PhD
  \cr Senior Medical Science Director
  \cr Genentech, Inc.
  \cr Evidence Science & Innovation | US Medical Affairs
  \cr \email{levy.drew@gene.com}

David C. Norris, MD
  \cr David Norris Consulting, LLC
  \cr \email{david@dnc-llc.com}
}
\seealso{
\code{\link[Hmisc:Overview]{Hmisc}}
  \code{\link[rms:rms]{rms}}
}
\keyword{package}

