\docType{data}
\name{vlbw}
\alias{vlbw}
\title{Data on 671 very low birth weight infants}
\format{A data frame containing 32 observations on 671 infants.}
\source{
  \url{http://biostat.mc.vanderbilt.edu/wiki/Main/DataSets}
}
\description{
  Documentation for included dataset \code{vlbw}
}
\details{
  Data on 671 infants with very low (<1600 grams) birth
  weight from 1981-87 were collected at Duke University
  Medical Center by Dr. Michael O'Shea.
}
\references{
  O'Shea M, Savitz DA, Hage ML, Feinstein KA: Prenatal
  events and the risk of subependymal / intraventricular
  haemorrhage in very low birth weight neonates.
  \emph{Paediatric and Perinatal Epidemiology}
  1992;6:352-362.
}
\keyword{datasets}

