% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbp_color.R
\name{cbp_pal}
\alias{cbp_pal}
\title{A colorblind proof color palette}
\usage{
cbp_pal(black = FALSE)
}
\arguments{
\item{black}{logical use the palette with grey or black color (default = FALSE)}
}
\description{
A colorblind proof color palette
}
\examples{
show_pal(cbp_pal(black = FALSE)(8))
show_pal(cbp_pal(black = TRUE)(8))

}
