\name{key}
\alias{key}
\title{Generic function 'key'}
\usage{
  key(x, ...)
}
\arguments{
  \item{x}{A colored object}

  \item{\dots}{Additional arguments}
}
\value{
  A vector of colors
}
\description{
  This generic function supports extraction of color keys
  from colored factors, via the specific method
  \code{key.colored}
}

