% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.plotbraids.R
\name{panel.plotbraids}
\alias{panel.plotbraids}
\title{panel.plotbraids}
\usage{
panel.plotbraids(x, y, subscripts, ..., formula, data, idvar, stratify, steps,
  color.by, outside)
}
\arguments{
\item{x}{}

\item{y}{}

\item{subscripts}{}

\item{...}{}

\item{formula}{}

\item{data}{}

\item{idvar}{}

\item{stratify}{}

\item{steps}{}

\item{color.by}{}
}
\value{
None
}
\description{
Panel function for function 'plotbraids'
}
\details{
This panel function manages details of braid geometry.  Key data
preparation steps are performed by the high-level 'plotbraids'
function.
}
\author{
David C. Norris
}
\seealso{
\code{\link{plotbraids}}
}
\keyword{hplot}
\keyword{internal}

